+++ test/setup.test.js
@@ -4,7 +4,7 @@ const chai = require('chai');
 const chaiHttp = require('chai-http');
 
 const { TEST_DATABASE_URL } = require('../config');
-const { dbConnect, dbDisconnect } = require('../db/db-mongoose');
+const { dbConnect, dbDisconnect } = require('../db/mongoose');
 // const {dbConnect, dbDisconnect} = require('../db-knex');
 
 // Set NODE_ENV to `test` to disable http layer logs
